Add files via upload
Sum.cpp  is a multithreaded program that simulates guessing the correct values for a combination lock using POSIX threads (Pthreads). The combo.txt file was on a different server and was lost, but essentially it held a series of numbers (representing the correct combination for a lock) from a file called combo.txt and stores them in an array named combo.