Add files via upload
This code is a multithreaded C++ program that uses Pthreads (POSIX threads) to create a number of threads (provided as a command-line argument) and has each thread print a message. The program also calculates how many threads successfully executed by incrementing a global sum variable. This program is good to understand the concept of multithreading and to expose the potential issue of race conditions in concurrent programming.
                          - (A race condition occurs when two or more threads try to modify shared data at the same time without proper synchronization. This leads to                               unpredictable or incorrect behavior, as the outcome depends on the order in which the threads access the shared resource.)